% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterise_recomb_landscape.R
\name{characterise_recomb_landscape}
\alias{characterise_recomb_landscape}
\title{characterise_recomb_landscape: Characterise recombination landscape using ordered linkage map and physical map data.}
\usage{
characterise_recomb_landscape(LG.vec, LG.pos, LG.cM, window.size,
  verbose = TRUE)
}
\arguments{
\item{LG.vec}{vector of chromosome/linkage map identifiers.}

\item{LG.pos}{vector of base pair positions.}

\item{LG.cM}{vector of linkage map positions.}

\item{window.size}{integer Window size in which recombination rate is estimated (base pairs).}

\item{verbose}{logical. Default is TRUE.}
}
\description{
characterise_recomb_landscape: Characterise recombination landscape using ordered linkage map and physical map data.
}
