% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_double_crossovers.R
\name{check_double_crossovers}
\alias{check_double_crossovers}
\title{check_double_crossovers: Split chromosomes into segments of runs of SNPs
inherited from particular grandparents. Used as quality control to determine
double crossovers over short genomic distances.}
\usage{
check_double_crossovers(parsed.xovers, physical.map = NULL)
}
\arguments{
\item{parsed.xovers}{Output from parse_crossovers}

\item{physical.map}{Optional data.frame. Must have columns as follows, or
will throw error: c("SNP.Name", "Position", "Order", "analysisID").}
}
\description{
check_double_crossovers: Split chromosomes into segments of runs of SNPs
inherited from particular grandparents. Used as quality control to determine
double crossovers over short genomic distances.
}
