% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_crimap_input.R
\name{create_crimap_input}
\alias{create_crimap_input}
\title{create_crimap_input: Create a CriMAP input file.}
\usage{
create_crimap_input(gwaa.data, familyPedigree, analysisID, snplist = NULL,
  chr = NULL, is.X = FALSE, is.Z = FALSE, pseudoautoSNPs = NULL,
  mend.errors = NULL, outdir = NULL, verbose = TRUE,
  clear.existing.analysisID = TRUE)
}
\arguments{
\item{gwaa.data}{GenABEL gwaa.data object containing genomic data}

\item{familyPedigree}{data.frame containing columns ANIMAL, FATHER, MOTHER
and FAMILY. FAMILY defines family groups for crimap. IDs in ANIMAL can be
repeated within and between families if necessary.}

\item{analysisID}{string. Used as an identifier for the chromosome. Must
begin with a numeric character as per crimap convention; for example,
analysisID = "1a" will output the file "chr1a.gen".}

\item{snplist}{vector, optional. A list of ordered SNPs. UNless build is run,
SNPs will be assumed to be in this order.}

\item{is.X}{logical. If true, then all heterozygotes are scored as missing in
males, except for pseudoautosomal SNPs when defined (see below)}

\item{is.Z}{logical. If true, then all heterozygotes are scored as missing in
females, except for pseudoautosomal SNPs when defined (see below)}

\item{pseudoautoSNPs}{Character vector of pseudoautosomal SNP IDs. Only used
if is.X or is.Z == TRUE.}

\item{mend.errors}{TO DO}

\item{outdir}{String, optional. Specify the path of the directory in which
the output file should be written.}

\item{verbose}{Logical. FALSE will suppress messages.}

\item{clear.existing.analysisID}{Logical, default is TRUE. Overwrites .gen
file and deletes crimap files that have the analysisID in the target
directory.}
}
\description{
create_crimap_input: Create a CriMAP input file.
}

