% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCrimapFile.R
\name{createCrimapFiles}
\alias{createCrimapFiles}
\title{genabelSubset: Subset a genabel dataset based on specified criteria.}
\usage{
createCrimapFiles(gwaa.data, familyPedigree, analysisID, snplist = NULL,
  chr = NULL, is.X = FALSE, is.Z = FALSE, pseudoautoSNPs = NULL,
  mend.errors = NULL, outdir = "crimap", verbose = TRUE)
}
\arguments{
\item{gwaa.data}{GenABEL gwaa.data to subset}

\item{snplist}{vector of ordered SNP loci.}

\item{chr}{vector, optional: retain SNPs on particular chromosome(s)}

\item{ped}{data.frame with pedigree.}
}
\description{
genabelSubset: Subset a genabel dataset based on specified criteria.
}

