% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deer.abel}
\alias{deer.abel}
\title{SNP genotype data from Red Deer (Cervus Elaphus) on the Isle of Rum, Scotland.}
\format{A GenABEL gwaa.data object with  frame with 6464 SNPs in 17 individuals.}
\usage{
deer.abel
}
\description{
SNP genotype data from Red Deer (Cervus Elaphus) on the Isle of Rum, Scotland.
}
\keyword{datasets}

