% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deer.famped}
\alias{deer.famped}
\title{CriMAP pedigrees of Red Deer (Cervus Elaphus) on the Isle of Rum, Scotland.}
\format{Pedigrees for CriMAP analysis with ANIMAL, FATHER, MOTHER and a Family identifier.}
\usage{
deer.famped
}
\description{
CriMAP pedigrees of Red Deer (Cervus Elaphus) on the Isle of Rum, Scotland.
}
\keyword{datasets}

