% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deer.ped}
\alias{deer.ped}
\title{Pedigree of 14 Red Deer (Cervus Elaphus) on the Isle of Rum, Scotland.}
\format{Pedigree in format ANIMAL, FATHER, MOTHER}
\usage{
deer.ped
}
\description{
Pedigree of 14 Red Deer (Cervus Elaphus) on the Isle of Rum, Scotland.
}
\keyword{datasets}

