% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{format_family_pedigree}
\alias{format_family_pedigree}
\title{format_family_pedigree: Format the family pedigree for constistency}
\usage{
format_family_pedigree(familyPedigree)
}
\arguments{
\item{familyPedigree}{data.frame containing columns ANIMAL, FATHER, MOTHER
and FAMILY. FAMILY defines family groups for crimap. IDs in ANIMAL can be
repeated within and between families if necessary.}
}
\description{
format_family_pedigree: Format the family pedigree for constistency
}

