% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genabel_subset.R
\name{genabel_subset}
\alias{genabel_subset}
\title{genabel_subset: Subset a genabel dataset based on specified criteria.}
\usage{
genabel_subset(gwaa.data, chr = NULL, ped = NULL, snplist = NULL)
}
\arguments{
\item{gwaa.data}{GenABEL gwaa.data to subset}

\item{chr}{vector, optional: retain SNPs on particular chromosome(s)}

\item{ped}{data.frame with pedigree.}

\item{snplist}{vector of ordered SNP loci.}
}
\description{
genabel_subset: Subset a genabel dataset based on specified criteria.
}
