% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genabelSubset.R
\name{genabelSubset}
\alias{genabelSubset}
\title{genabelSubset: Subset a genabel dataset based on specified criteria.}
\usage{
genabelSubset(gwaa.data, chr = NULL, ped = NULL, snplist = NULL)
}
\arguments{
\item{gwaa.data}{GenABEL gwaa.data to subset}

\item{chr}{vector, optional: retain SNPs on particular chromosome(s)}

\item{ped}{data.frame with pedigree.}

\item{snplist}{vector of ordered SNP loci.}
}
\description{
genabelSubset: Subset a genabel dataset based on specified criteria.
}

