% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{InfLengthFunc}
\alias{InfLengthFunc}
\title{InfLengthFunc: Determine positions of first and last informative loci from chrompic output}
\usage{
InfLengthFunc(CMPstring, position = "First")
}
\arguments{
\item{CMPstring}{String of crossover events and informative loci from chrompic output}

\item{position}{"First" or "Last", return either the first or last informative locus position.}
}
\description{
InfLengthFunc: Determine positions of first and last informative loci from chrompic output
}
