% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_build_insert.R
\name{parse_built_insert}
\alias{parse_built_insert}
\title{parse_build_insert: Parse crossover information from CriMAP chrompic output}
\usage{
parse_built_insert(buildfile, snpinsert)
}
\arguments{
\item{buildfile}{File with output from build with inserted loci}

\item{snpinsert}{Vector of SNPs that were inserted in prepare and build}
}
\description{
parse_build_insert: Parse crossover information from CriMAP chrompic output
}
