% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_crossovers.R
\name{parse_crossovers}
\alias{parse_crossovers}
\title{parse_crossovers: Parse crossover information from CriMAP chrompic output}
\usage{
parse_crossovers(chrompicfile)
}
\arguments{
\item{chrompicfile}{File with output from chrompic}
}
\description{
parse_crossovers: Parse crossover information from CriMAP chrompic output
}

