% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_flips.R
\name{parse_flips}
\alias{parse_flips}
\title{parse_flips: Parse output from CriMAP flips}
\usage{
parse_flips(flipsfile)
}
\arguments{
\item{flipsfile}{File with output from flips.}
}
\description{
parse_flips: Parse output from CriMAP flips
}
