% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_loc.R
\name{parse_loc}
\alias{parse_loc}
\title{parse_loc: Parse output from CriMAP .loc files}
\usage{
parse_loc(locfile)
}
\arguments{
\item{locfile}{.loc file}
}
\description{
parse_loc: Parse output from CriMAP .loc files
}

