% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_map.R
\name{parse_map}
\alias{parse_map}
\title{parse_map: Parse map from CriMAP map output}
\usage{
parse_map(mapfile)
}
\arguments{
\item{mapfile}{File with output from map}
}
\description{
parse_map: Parse map from CriMAP map output
}
