% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_map_build.R
\name{parse_map_build}
\alias{parse_map_build}
\title{parse_map_build: Parse map from CriMAP chrompic output}
\usage{
parse_map_build(buildfile)
}
\arguments{
\item{buildfile}{File with output from build}
}
\description{
parse_map_build: Parse map from CriMAP chrompic output
}

