% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_map_chrompic.R
\name{parse_map_chrompic}
\alias{parse_map_chrompic}
\title{parse_map_chrompic: Parse map from CriMAP chrompic output}
\usage{
parse_map_chrompic(chrompicfile)
}
\arguments{
\item{chrompicfile}{File with output from chrompic}
}
\description{
parse_map_chrompic: Parse map from CriMAP chrompic output
}

