% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseCrossovers.R
\name{parseCrossovers}
\alias{parseCrossovers}
\title{parseCrossovers: Parse crossover information from CriMAP chrompic output}
\usage{
parseCrossovers(chrompicfile)
}
\arguments{
\item{chrompicfile}{File with output from chrompic}
}
\description{
parseCrossovers: Parse crossover information from CriMAP chrompic output
}

