% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFlips.R
\name{parseFlips}
\alias{parseFlips}
\title{parseFlips: Parse output from CriMAP flips}
\usage{
parseFlips(flipsfile)
}
\arguments{
\item{flipsfile}{File with output from flips.}
}
\description{
parseFlips: Parse output from CriMAP flips
}

