% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMap.R
\name{parseMap}
\alias{parseMap}
\title{parseMap: Parse map from CriMAP map output}
\usage{
parseMap(mapfile)
}
\arguments{
\item{mapfile}{File with output from map}
}
\description{
parseMap: Parse map from CriMAP map output
}

