% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseMapChrompic.R
\name{parseMapChrompic}
\alias{parseMapChrompic}
\title{parseMapChrompic: Parse map from CriMAP chrompic output}
\usage{
parseMapChrompic(chrompicfile)
}
\arguments{
\item{chrompicfile}{File with output from chrompic}
}
\description{
parseMapChrompic: Parse map from CriMAP chrompic output
}

