% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree.format.R
\name{pedigree.format}
\alias{pedigree.format}
\title{Format Pedigree
This function formats the pedigree for downstream analysis.}
\usage{
pedigree.format(ped, pedigree.type = "simple")
}
\arguments{
\item{ped}{Pedigree object in "simple" format (Three columns for ANIMAL,
MOTHER and FATHER) or in "plink" format (Five to Six columns for FAMILY,
ANIMAL, FATHER, MOTHER, SEX and Phenotype, where the phenotype column is
optional). The simple argument can recognise the order of parents if
they are named sensibly. Run simple.ped.name.rules() for an example.}

\item{pedigree.type}{Defaults to "simple", can also accept "plink" which
is equivalent for for first 5 to 6 columns of a PLINK .ped file.}
}
\description{
Format Pedigree
This function formats the pedigree for downstream analysis.
}

