% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{RecCountFunc}
\alias{RecCountFunc}
\title{RecCountFunc: Determine recombination count from chrompic output strings}
\usage{
RecCountFunc(CMPstring.vec)
}
\arguments{
\item{CMPstring.vec}{vector of strings of crossover events and informative
loci from chrompic output}
}
\description{
RecCountFunc: Determine recombination count from chrompic output strings
}
