% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revise_double_crossovers.R
\name{revise_double_crossovers}
\alias{revise_double_crossovers}
\title{revise_double_crossovers: remove erroneously phased loci from parsed
crossover data.}
\usage{
revise_double_crossovers(parsed.xovers, removesections)
}
\arguments{
\item{parsed.xovers}{output from parse_crossovers}

\item{removesections}{data.frame that contains the following fields: StartPos, StopPos and
UniqueID from check_double_crossovers function. Frame should only include
values that should be removed}
}
\description{
revise_double_crossovers: remove erroneously phased loci from parsed
crossover data.
}
