% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_crimap_build.R
\name{run_crimap_build}
\alias{run_crimap_build}
\title{run_crimap_build: Run chrompic in crimap}
\usage{
run_crimap_build(genfile, PLT = NULL)
}
\arguments{
\item{genfile}{path to the .gen file for running the chrompic function.}

\item{PLT}{numeric. Optional. Specify PUK_LIKE_TOL and PK_LIKE_TOL value.}
}
\description{
run_crimap_build: Run chrompic in crimap
}

