% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_crimap_chrompic.R
\name{run_crimap_chrompic}
\alias{run_crimap_chrompic}
\title{run_crimap_chrompic: Run prepare in crimap}
\usage{
run_crimap_chrompic(crimap.path, genfile)
}
\arguments{
\item{crimap.path}{path to run crimap. This should be given relative to the
same directory as...}

\item{genfile}{path to the .gen file for running the chrompic function.}
}
\description{
NB. This function will only work if the crimap file is in a sub-directory of
the working directory, and that the crimap.path is in the same directory as
the crimap file. This function will delete all files associated with the
analysis suffix, except the .gen file.
}

