% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_crimap_map.R
\name{run_crimap_map}
\alias{run_crimap_map}
\title{run_crimap_map: Run prepare in crimap}
\usage{
run_crimap_map(genfile)
}
\arguments{
\item{genfile}{path to the .gen file for running the chrompic function.}
}
\description{
run_crimap_map: Run prepare in crimap
}

