% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_crimap_prepare.R
\name{run_crimap_prepare}
\alias{run_crimap_prepare}
\title{run_crimap_prepare: Run prepare in crimap}
\usage{
run_crimap_prepare(genfile, build = FALSE)
}
\arguments{
\item{genfile}{path for crimap .gen file.}
}
\description{
NB. This function will delete all files associated with the
analysis suffix, except the .gen file.
}

