% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_crimap_twopoint.R
\name{run_crimap_twopoint}
\alias{run_crimap_twopoint}
\title{run_crimap_twopoint: Run twopoint in crimap}
\usage{
run_crimap_twopoint(genfile)
}
\arguments{
\item{genfile}{path to the .gen file for running the twopoint function.}
}
\description{
run_crimap_twopoint: Run twopoint in crimap
}
