% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_functions.R
\name{SingletonStatusFunc}
\alias{SingletonStatusFunc}
\title{SingletonStatusFunc: Flag if there is a double crossover spanning a single SNP}
\usage{
SingletonStatusFunc(CMPstring.vec)
}
\arguments{
\item{CMPstring.vec}{vector of strings of crossover events and informative
loci from chrompic output}
}
\description{
SingletonStatusFunc: Flag if there is a double crossover spanning a single SNP
}
