% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_position_cmpstring.R
\name{switch_position_cmpstring}
\alias{switch_position_cmpstring}
\title{switch_position_cmpstring: function to pull out chromosome segments of
different grandparental origin. Used within the function
check_double_crossovers}
\usage{
switch_position_cmpstring(CMPstring, UniqueID = NULL)
}
\arguments{
\item{CMPstring}{a chrompic strings}

\item{UniqueID}{unique identifier}
}
\description{
switch_position_cmpstring: function to pull out chromosome segments of
different grandparental origin. Used within the function
check_double_crossovers
}
