% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_crimap_chrompic.R
\name{run_crimap_chrompic}
\alias{run_crimap_chrompic}
\title{run_crimap_chrompic: Run chrompic in crimap}
\usage{
run_crimap_chrompic(genfile)
}
\arguments{
\item{genfile}{path to the .gen file for running the chrompic function.}
}
\description{
run_crimap_chrompic: Run chrompic in crimap
}
