% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_crimap_flips.R
\name{run_crimap_flips}
\alias{run_crimap_flips}
\title{run_crimap_flips: Run flips}
\usage{
run_crimap_flips(crimap.path, genfile, flips = 2)
}
\arguments{
\item{crimap.path}{path to run crimap. This should be given relative to the
same directory as...}

\item{genfile}{path to the .gen file for running the chrompic function.}

\item{flips}{number of loci to flip.}
}
\description{
run_crimap_flips: Run flips
}

