% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_crimap_prepare.R
\name{run_crimap_prepare}
\alias{run_crimap_prepare}
\title{run_crimap_prepare: Run prepare in crimap}
\usage{
run_crimap_prepare(crimap.path, crimap.file)
}
\arguments{
\item{crimap.path}{path to run crimap. This should be given relative to the same directory
as...}

\item{crimap.file}{path for crimap file. Should be in a subdirectory of the
working directory.}
}
\description{
NB. This function will only work if the crimap file is in a sub-directory of
the working directory, and that the crimap.path is in the same directory as
the crimap file. This function will delete all files associated with the
analysis suffix, except the .gen file.
}

