% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_crimap_prepare.R
\name{run_crimap_prepare}
\alias{run_crimap_prepare}
\title{run_crimap_prepare: Run prepare in crimap}
\usage{
run_crimap_prepare(genfile, build = FALSE, snplist = NULL,
  snpinsert = NULL)
}
\arguments{
\item{genfile}{path for crimap .gen file.}

\item{build}{logical. Tells crimap whether to prepare for build analysis.}

\item{snplist}{character vector. Ordered loci for build analysis.}

\item{snpinset}{character vector. Loci to be added to the build.}
}
\description{
NB. This function will delete all files associated with the
analysis suffix, except the .gen file.
}

